requirements.txt
langchain
langchain-community
langchain-chroma
langchain-openai
chromadb
sentence-transformers
streamlit
pymupdf
pandas
python-dotenv
pydantic
pydantic-settings
crewai
docker
langchain-huggingface

langgraph
openai
tabulate
openpyxl

